+++ tests/testthat/test_add_locked_in_constraints.R
@@ -100,6 +100,22 @@ test_that("character locked in data (solve)", {
   expect_true(all(s$solution_1[which(sim_pu_polygons$locked_in)] == 1))
 })
 
+test_that("character locked in data with proportion decisions (solve)", {
+  skip_on_cran()
+  skip_if_not(any_solvers_installed())
+  # create problem
+  data(sim_pu_polygons, sim_features)
+  s <- problem(sim_pu_polygons, sim_features, "cost") %>%
+    add_min_set_objective() %>%
+    add_relative_targets(0.1) %>%
+    add_proportion_decisions() %>%
+    add_locked_in_constraints("locked_in") %>%
+    add_default_solver(time_limit = 5) %>%
+    solve()
+  # check that the solution obeys constraints as expected
+  expect_true(all(s$solution_1[which(sim_pu_polygons$locked_in)] == 1))
+})
+
 test_that("raster locked in data (compile)", {
   # create problem
   data(sim_pu_raster, sim_locked_in_raster, sim_features)